move password reset settings to "instance" section
