+++ web/messages/en/settings.json
@@ -11,6 +11,8 @@
   "settings_instance_section_core_description": "Configure the primary instance identity, URLs and session lifetime used across the platform.",
   "settings_instance_section_data_retention": "VPN stats retention",
   "settings_instance_section_data_retention_description": "Control if VPN statistics are purged automatically and how long they are kept.",
+  "settings_instance_section_password_title": "Password settings",
+  "settings_instance_section_password_description": "Set the token expiration time when a password is changed in the system.",
   "settings_instance_error_invalid_url": "Invalid URL",
   "settings_instance_error_invalid_host": "Defguard URL must use a hostname, not an IP address",
   "settings_instance_error_defguard_url_required": "Defguard URL is required",
@@ -18,6 +20,7 @@
   "settings_instance_label_name": "Instance name",
   "settings_instance_label_public_proxy_url": "Public Edge Component URL",
   "settings_instance_label_session_duration": "Session duration",
+  "settings_instance_label_password_reset_session_expiration": "Password reset session expiration",
   "settings_vpn_stats_toggle_title": "Stats purge",
   "settings_vpn_stats_label_purge_frequency": "Stats purge frequency",
   "settings_vpn_stats_label_purge_threshold": "Stats purge threshold",
@@ -65,7 +68,6 @@
   "settings_enrollment_toggle_reset_password_title": "Display the \"Reset password\" option on the Enrollment Service home screen.",
   "settings_enrollment_toggle_reset_password_description": "In case your instance uses only an external SSO provider, Defguard's internal password reset is not needed, as authentication is handled externally.",
   "settings_general_section_instance_content": "Configure your instance name and branding settings. Add a logo to personalize the interface and make it easily recognizable to your users.",
-  "settings_general_section_password_reset_content": "Configure password reset token and session timeouts for self-service password recovery.",
   "settings_general_section_client_behavior_content": "Manage how users interact with the Defguard client. Control device management permissions, configuration access, and traffic routing options.",
   "settings_general_section_enrollment_content": "Configure enrollment settings to define how users are invited, registered, and onboarded into your instance.",
   "settings_duration_one_day": "1 day",
