--- web/src/pages/settings/SettingsEnrollmentPage/SettingsEnrollmentPage.tsx
@@ -1,194 +1,0 @@
-import { useMutation, useQuery } from '@tanstack/react-query';
-import { Link } from '@tanstack/react-router';
-import { useMemo } from 'react';
-import z from 'zod';
-import { m } from '../../../paraglide/messages';
-import api from '../../../shared/api/api';
-import type { Settings } from '../../../shared/api/types';
-import { Breadcrumbs } from '../../../shared/components/Breadcrumbs/Breadcrumbs';
-import { Controls } from '../../../shared/components/Controls/Controls';
-import { Page } from '../../../shared/components/Page/Page';
-import { SettingsCard } from '../../../shared/components/SettingsCard/SettingsCard';
-import { SettingsHeader } from '../../../shared/components/SettingsHeader/SettingsHeader';
-import { SettingsLayout } from '../../../shared/components/SettingsLayout/SettingsLayout';
-import { Button } from '../../../shared/defguard-ui/components/Button/Button';
-import { SizedBox } from '../../../shared/defguard-ui/components/SizedBox/SizedBox';
-import { Snackbar } from '../../../shared/defguard-ui/providers/snackbar/snackbar';
-import { ThemeSpacing } from '../../../shared/defguard-ui/types';
-import { isPresent } from '../../../shared/defguard-ui/utils/isPresent';
-import { useAppForm } from '../../../shared/form';
-import { formChangeLogic } from '../../../shared/formLogic';
-import { getSettingsQueryOptions } from '../../../shared/query';
-import {
-  createNumericSelectOptions,
-  withNumericFallbackOption,
-} from '../../../shared/utils/numericSelectOptions';
-
-const breadcrumbs = [
-  <Link
-    to="/settings"
-    search={{
-      tab: 'general',
-    }}
-    key={0}
-  >
-    {m.settings_breadcrumb_general()}
-  </Link>,
-  <Link to="/settings/enrollment" key={1}>
-    {m.settings_breadcrumb_password_reset()}
-  </Link>,
-];
-
-export const SettingsEnrollmentPage = () => {
-  const { data: settings } = useQuery(getSettingsQueryOptions);
-  return (
-    <Page title={m.settings_page_title()}>
-      <Breadcrumbs links={breadcrumbs} />
-      <SettingsLayout>
-        <SettingsHeader
-          icon="customize"
-          title={m.settings_password_reset_title()}
-          subtitle={m.settings_password_reset_subtitle()}
-        />
-        {isPresent(settings) && (
-          <SettingsCard>
-            <Content settings={settings} />
-          </SettingsCard>
-        )}
-      </SettingsLayout>
-    </Page>
-  );
-};
-
-const formSchema = z.object({
-  password_reset_token_timeout_hours: z.number(m.form_error_required()).int().min(1),
-  password_reset_session_timeout_minutes: z.number(m.form_error_required()).int().min(1),
-});
-
-type FormFields = z.infer<typeof formSchema>;
-
-const passwordResetTokenTimeoutBaseOptions = createNumericSelectOptions({
-  1: m.settings_duration_one_hour(),
-  12: m.settings_duration_hours({ hours: 12 }),
-  24: m.settings_duration_one_day(),
-  168: m.settings_duration_one_week(),
-});
-
-const passwordResetSessionTimeoutBaseOptions = createNumericSelectOptions({
-  10: m.settings_duration_minutes({ minutes: 10 }),
-  30: m.settings_duration_minutes({ minutes: 30 }),
-  60: m.settings_duration_one_hour(),
-});
-
-const Content = ({ settings }: { settings: Settings }) => {
-  const { mutateAsync } = useMutation({
-    mutationFn: api.settings.patchSettings,
-    meta: {
-      invalidate: ['settings'],
-    },
-    onSuccess: () => {
-      Snackbar.default(m.settings_msg_saved());
-    },
-    onError: () => {
-      Snackbar.error(m.settings_msg_save_failed());
-    },
-  });
-
-  const defaultValues = useMemo(
-    (): FormFields => ({
-      password_reset_token_timeout_hours:
-        settings.password_reset_token_timeout_hours ?? 24,
-      password_reset_session_timeout_minutes:
-        settings.password_reset_session_timeout_minutes ?? 10,
-    }),
-    [settings],
-  );
-
-  const passwordResetTokenTimeoutOptions = useMemo(
-    () =>
-      withNumericFallbackOption(
-        passwordResetTokenTimeoutBaseOptions,
-        defaultValues.password_reset_token_timeout_hours,
-        'hours',
-      ),
-    [defaultValues.password_reset_token_timeout_hours],
-  );
-
-  const passwordResetSessionTimeoutOptions = useMemo(
-    () =>
-      withNumericFallbackOption(
-        passwordResetSessionTimeoutBaseOptions,
-        defaultValues.password_reset_session_timeout_minutes,
-        'minutes',
-      ),
-    [defaultValues.password_reset_session_timeout_minutes],
-  );
-
-  const form = useAppForm({
-    defaultValues,
-    validationLogic: formChangeLogic,
-    validators: {
-      onSubmit: formSchema,
-      onChange: formSchema,
-    },
-    onSubmit: async ({ value }) => {
-      await mutateAsync(value);
-      form.reset(value);
-    },
-  });
-
-  return (
-    <form
-      onSubmit={(e) => {
-        e.stopPropagation();
-        e.preventDefault();
-        form.handleSubmit();
-      }}
-    >
-      <form.AppForm>
-        <form.AppField name="password_reset_token_timeout_hours">
-          {(field) => (
-            <field.FormSelect
-              required
-              label={m.settings_enrollment_label_password_reset_token_validity()}
-              options={passwordResetTokenTimeoutOptions}
-            />
-          )}
-        </form.AppField>
-        <SizedBox height={ThemeSpacing.Xl} />
-        <form.AppField name="password_reset_session_timeout_minutes">
-          {(field) => (
-            <field.FormSelect
-              required
-              label={m.settings_enrollment_label_password_reset_session_expires_in()}
-              options={passwordResetSessionTimeoutOptions}
-            />
-          )}
-        </form.AppField>
-      </form.AppForm>
-      <form.Subscribe
-        selector={(s) => ({
-          isDefault: s.isDefaultValue || s.isPristine,
-          isSubmitting: s.isSubmitting,
-        })}
-      >
-        {({ isDefault, isSubmitting }) => (
-          <Controls>
-            <div className="right">
-              <Button
-                variant="primary"
-                text={m.controls_save_changes()}
-                disabled={isDefault}
-                loading={isSubmitting}
-                type="submit"
-                onClick={() => {
-                  form.handleSubmit();
-                }}
-              />
-            </div>
-          </Controls>
-        )}
-      </form.Subscribe>
-    </form>
-  );
-};
